requirements.txt
# FastAPI and ASGI Server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Pydantic
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
pyjwt==2.8.1
python-multipart==0.0.6

# Redis
redis==5.0.1

# HTTP Client
httpx==0.25.2

# Content Sanitization
bleach==6.1.0
markdown==3.5.1

# Rate Limiting
slowapi==0.1.9

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Code Quality
black==23.12.0
ruff==0.1.8
mypy==1.7.1
bandit==1.7.5

# Security Scanning
pip-audit==2.6.1
safety==2.3.5
